requirements.txt
fastapi
uvicorn
python-dotenv
pymongo
langchain
langchain-community
langchain-google-genai
pdfplumber
pandas
paddleocr
PyMuPDF